refactor replacer to use extractor
