+++ src/Replacer/Extractor/Tokenizer.php
@@ -4,26 +4,25 @@ declare(strict_types=1);
 
 // @php-cs-fixer-ignore array_indentation
 
-namespace VStelmakh\UrlHighlight\Replacer\Tokenizer;
+namespace VStelmakh\UrlHighlight\Replacer\Extractor;
 
-use VStelmakh\UrlHighlight\Replacer\Tokenizer\Token\CommentToken;
-use VStelmakh\UrlHighlight\Replacer\Tokenizer\Token\PlainToken;
-use VStelmakh\UrlHighlight\Replacer\Tokenizer\Token\TagToken;
-use VStelmakh\UrlHighlight\Replacer\Tokenizer\Token\TagType;
-use VStelmakh\UrlHighlight\Replacer\Tokenizer\Token\Token;
+use VStelmakh\UrlHighlight\Replacer\Extractor\Token\CommentToken;
+use VStelmakh\UrlHighlight\Replacer\Extractor\Token\PlainToken;
+use VStelmakh\UrlHighlight\Replacer\Extractor\Token\TagToken;
+use VStelmakh\UrlHighlight\Replacer\Extractor\Token\TagType;
+use VStelmakh\UrlHighlight\Replacer\Extractor\Token\Token;
 
 /**
  * Splits an HTML string into a sequence of {@see Token}.
  *
  * @internal
  */
-final readonly class HtmlTokenizer implements Tokenizer
+final readonly class Tokenizer
 {
     /**
      * @return \Generator<Token>
      */
-    #[\Override]
-    public function tokenize(string $text): \Generator
+    public function tokenize(string $html): \Generator
     {
         // Matches HTML comments and tags, handling quoted attributes so that ">" inside values is not treated as tag.
         $pattern = implode('', [
@@ -46,11 +45,11 @@ final readonly class HtmlTokenizer implements Tokenizer
 
         // Matching one tag at a time from a moving cursor, rather than splitting the whole input up front, keeps only
         // the current token in memory. This significantly saves memory for markup-dense inputs.
-        while (preg_match($pattern, $text, $match, PREG_OFFSET_CAPTURE, $cursor) === 1) {
+        while (preg_match($pattern, $html, $match, PREG_OFFSET_CAPTURE, $cursor) === 1) {
             [$contents, $contentsOffset] = $match[0];
 
             // Adjacent tags, e.g. "</b><i>", leave nothing in between.
-            $plain = substr($text, $cursor, $contentsOffset - $cursor);
+            $plain = substr($html, $cursor, $contentsOffset - $cursor);
             if ($plain !== '') {
                 yield new PlainToken($plain);
             }
@@ -61,7 +60,7 @@ final readonly class HtmlTokenizer implements Tokenizer
         }
 
         // On PCRE failure the rest is treated as plain text.
-        yield new PlainToken(substr($text, $cursor));
+        yield new PlainToken(substr($html, $cursor));
     }
 
     private function createTagToken(string $contents): TagToken
